+++ src/main/kotlin/klib/extensions/MutableMap.kt
@@ -22,7 +22,7 @@ fun <K, V> MutableMap<K, V>.smartMerge(other: Map<K, V>) {
  * @param other The map to merge
  * @param convertIt Converter function to put Any to type 2
  *
- * @since 3.0.0
+ * @since 3.1.0
  * @author Thomas Obernosterer
  */
 fun <K, V> MutableMap<K, V>.smartMerge(other: Map<K, Any>, convertIt: (Any) -> V) {
@@ -53,6 +53,26 @@ fun <K, V> MutableMap<K, V>.mergeArrays(keys: Array<K>, values: Array<V>) {
 }
 
 /**
+ * Merge a Key and a Value Array into a MutableMap; Duplicate keys will be ignored
+ *
+ * @param keys The Keys
+ * @param values The Values
+ * @param convertIt Converter function to put Any to type 2
+ *
+ * @since 3.1.0
+ * @author Thomas Obernosterer
+ */
+fun <K, V> MutableMap<K, V>.mergeArrays(keys: Array<K>, values: Array<Any>, convertIt: (Any) -> V) {
+    if (keys.size == values.size) {
+        for (i in 0 until keys.size) {
+            if (keys[i] !in this) {
+                this[keys[i]] = convertIt(values[i])
+            }
+        }
+    }
+}
+
+/**
  * Merge a Map into a MutableMap with overwriting existing keys
  *
  * @param other The map to merge
@@ -63,5 +83,20 @@ fun <K, V> MutableMap<K, V>.mergeArrays(keys: Array<K>, values: Array<V>) {
 fun <K, V> MutableMap<K, V>.fullMerge(other: Map<K, V>) {
     other.forEach { (key, value) ->
         this[key] = value
+    }
+}
+
+/**
+ * Merge a Map into a MutableMap with overwriting existing keys
+ *
+ * @param other The map to merge
+ * @param convertIt Converter function to put Any to type 2
+ *
+ * @since 3.1.0
+ * @author Thomas Obernosterer
+ */
+fun <K, V> MutableMap<K, V>.fullMerge(other: Map<K, Any>, convertIt: (Any) -> V) {
+    other.forEach { (key, value) ->
+        this[key] = convertIt(value)
     }
 }