+++ src/main/kotlin/klib/kLibInf.kt
@@ -25,7 +25,7 @@ object kLibInf {
      * @since 2.0.0
      * @author Thomas Obernosterer
      */
-    val semver = SemVer(3, 0, 0)
+    val semver = SemVer(3, 1, 0)
 
     /**
      * Company creating kLib