+++ src/test/kotlin/devtests/MutableListToList.kt
@@ -1,0 +1,10 @@
+package devtests
+
+import klib.extensions.toListWithConvert
+
+fun main() {
+    val listA = mutableListOf("1","3","8")
+
+    val listB = listA.toListWithConvert { it.toIntOrNull() ?: 0 }
+    println(listB)
+}